OdooProjetMax/Prog-Odoo-1-main/tasks/requirements.txt
cookiecutter==1.6.0
future==0.16.0

invoke==1.1.1
# 3.11 was incompatible with kaptan for git-autoshare
PyYAML==3.13
requests==2.19.1
ruamel.yaml
gnupg==2.3.1
psycopg2-binary
python-dateutil==2.7.3
git-aggregator >= 1.6.0

git-autoshare >= 1.0.0b2


clipboard==0.0.4
